requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-WTF==1.2.1
Flask-Mail==0.9.1
SQLAlchemy==2.0.23
Werkzeug==2.3.7
Jinja2==3.1.2
WTForms==3.0.1
email-validator==2.1.0.post1
python-dotenv==1.0.0
Pillow==10.1.0
pyodbc==5.0.1
openai==1.3.5
Markdown==3.5.1
bleach==6.1.0
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0
coverage==7.3.2
